requirements.txt
flask
openai
python-dotenv
werkzeug
